Add text input with hardcoded validation (#1)
* Add text input with hardcoded validation

This adds an input field with a submit button. On submit the input is
validated against a hardcoded value and the border color of the input
field is updated. Currently this is pretty dumb and just shows different
versions of the input field based on state.

This also adds KeyboardResponder which return the height of the
keyboard. It's a simple way to ensure the input is showing but it
doesn't adjust for where the input is on the screen.

* Update project settings (add KeyboardResponder and necessary stuff for running on my phone)

* Minor refactoring

Co-authored-by: Steven Loria <sloria1@gmail.com>